Split commands into different files & add bitcount
